+++ CHANGELOG.md
@@ -2,6 +2,7 @@
  - Updated dependency of react-native-video to 4.4.0
  - Added 'close'-Button on Fullscreen for Android
  - Added duration/play time for IOS (with upgrading react-native-video)
+ - Fixed prop types issue on TouchableOpacity causing this package not to work on React Native v0.62.
 
 ## 0.9.1
  - Fixed video repeating when loop turned off (thanks @mattvot)